+++ src/pip_license_checker/filters.clj
@@ -2,8 +2,11 @@
   "Filters for requirements"
   (:gen-class)
   (:require
-   [pip-license-checker.version :as version]
-   [clojure.string :as str]))
+   [clojure.spec.alpha :as s]
+   [clojure.string :as str]
+   ;;[clojure.spec.test.alpha :refer [instrument]]
+   [pip-license-checker.spec :as sp]
+   [pip-license-checker.version :as version]))
 
 ;; Skip line with -r /--requirement/-e etc, URLs, blank lines, comments
 (def regex-skip-line-internal #"(\s*(?:https?:\/\/|#|-).*)|(^\s*$)")
@@ -15,10 +18,18 @@
 (def regex-remove-wildcard #"\.\*")
 (def regex-split-specifier-ops #"(===|==|~=|!=|>=|<=|<|>)")
 
+(s/fdef remove-requirements-internal-rules
+  :args (s/cat :requirements ::sp/requirements)
+  :ret ::sp/requirements)
+
 (defn remove-requirements-internal-rules
   "Exclude requirements from sequence according to app's internal rules"
   [requirements]
   (remove #(re-matches regex-skip-line-internal %) requirements))
+
+(s/fdef remove-requirements-user-rules
+  :args (s/cat :pattern ::sp/opt-pattern :requirements ::sp/requirements)
+  :ret ::sp/requirements)
 
 (defn remove-requirements-user-rules
   "Exclude requirements from sequence according to user-defined pattern"
@@ -47,3 +58,10 @@
         specifiers (if (= specifiers-vec [nil]) nil specifiers-vec)
         result {:name package-name :specifiers specifiers}]
     result))
+
+;;
+;; Instrumented functions - uncomment only while testing
+;;
+
+;; (instrument `remove-requirements-internal-rules)
+;; (instrument `remove-requirements-user-rules)